requirements.txt
websocket-client
requests
flask
gunicorn